refactor: extract session domain types
