+++ src/domain.rs
@@ -1,2 +1,3 @@
 pub(crate) mod activity;
 pub(crate) mod elicitation;
+pub(crate) mod session;
