+++ src/protocol.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+pub(crate) use crate::domain::session::{SessionGroup, SessionSummary};
+
 // --- Client → Server messages ---
 
 #[derive(Debug, Serialize)]
@@ -561,63 +563,6 @@ pub struct SessionChildrenData {
     pub next_cursor: Option<String>,
     #[serde(default)]
     pub total_count: Option<u64>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-pub struct SessionGroup {
-    pub cwd: Option<String>,
-    #[serde(default)]
-    pub sessions: Vec<SessionSummary>,
-    /// ISO 8601 timestamp of the most recent activity in this group.
-    #[serde(default)]
-    pub latest_activity: Option<String>,
-    #[serde(default)]
-    pub total_count: Option<u64>,
-    #[serde(default)]
-    pub next_cursor: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-pub struct SessionSummary {
-    pub session_id: String,
-    #[serde(default)]
-    pub name: Option<String>,
-    #[serde(default)]
-    pub title: Option<String>,
-    /// Working directory for this session (may differ from group cwd for remote sessions).
-    #[serde(default)]
-    pub cwd: Option<String>,
-    #[serde(default)]
-    pub created_at: Option<String>,
-    #[serde(default)]
-    pub updated_at: Option<String>,
-    /// Parent session ID if this is a forked session.
-    #[serde(default)]
-    pub parent_session_id: Option<String>,
-    #[serde(default)]
-    pub fork_origin: Option<String>,
-    #[serde(default)]
-    pub session_kind: Option<String>,
-    /// Whether this session has child (forked) sessions.
-    #[serde(default)]
-    pub has_children: bool,
-    /// Number of direct forked child sessions.
-    #[serde(default)]
-    pub fork_count: u64,
-    #[serde(default)]
-    pub children: Vec<SessionSummary>,
-    #[serde(default)]
-    pub children_next_cursor: Option<String>,
-    #[serde(default)]
-    pub children_total_count: Option<u64>,
-    #[serde(default)]
-    pub node: Option<String>,
-    #[serde(default)]
-    pub node_id: Option<String>,
-    #[serde(default)]
-    pub attached: Option<bool>,
-    #[serde(default)]
-    pub runtime_state: Option<String>,
 }
 
 #[derive(Debug, Deserialize)]
